Broke up style-controls into different components
