+++ src/InlineStyleControls.jsx
@@ -1,0 +1,36 @@
+// React imports
+import React from 'react';
+
+// Application Components: 
+import StyleButton from './StyleButton'
+
+// Styling:
+import './InlineStyleControls.css';
+
+// Inlince Style Components
+const INLINE_STYLES = [
+  {label: 'Bold', style: 'BOLD'},
+  {label: 'Italic', style: 'ITALIC'},
+  {label: 'Underline', style: 'UNDERLINE'},
+  {label: 'Monospace', style: 'CODE'},
+];
+
+const InlineStyleControls = (props) => {
+  const currentStyle = props.editorState.getCurrentInlineStyle();
+  return (
+    <div className="RichEditor-controls">
+      {INLINE_STYLES.map(type =>
+        <StyleButton
+          key={type.label}
+          active={currentStyle.has(type.style)}
+          label={type.label}
+          onToggle={props.onToggle}
+          style={type.style}
+        />
+      )}
+    </div>
+  );
+};
+
+
+export default InlineStyleControls;
