+++ src/StyleButton.jsx
@@ -1,0 +1,31 @@
+// React imports
+import React, { Component } from 'react';
+
+// Styling:
+import './StyleButton.css';
+
+// Style button components
+class StyleButton extends Component {
+  constructor() {
+    super();
+    this.onToggle = (e) => {
+      e.preventDefault();
+      this.props.onToggle(this.props.style);
+    };
+  }
+
+  render() {
+    let className = 'RichEditor-styleButton';
+    if (this.props.active) {
+      className += ' RichEditor-activeButton';
+    }
+
+    return (
+      <span className={className} onMouseDown={this.onToggle}>
+        {this.props.label}
+      </span>
+    );
+  }
+}
+
+export default StyleButton;
